Re-jig attributes and values
Make attributes a temporary thing.  Store the mapping from attribute
name to attribute value in the abbreviation, not the entry, making entry
just have an array of Value, and we find the location in the array for a
specific attribute by looking in the abbreviation's mapping. This should
save a lot of space.